Organize tools, remove binaries
While I build on both Windows MinGW and Linux, there are things here
that are catered to Windows-only, so they are gone.

Utilities, scripts and such are added or moved to util/

I don't care about having builds for any of the targets that aren't of
use to me right now, so they are gone. I may build them later but there
are several compile-time options in my fork and I don't know how I want
to handle this yet.