Refactored register code
There was a mix of GET_REG/SET_REG and struct-based accesses to
registers.

Changed GET/SET_REG to REG_GET/SET and kept it that way for GPIOs and a
couple of other registers. This will certainly change at a later date.

Changed RCC and FLASH to struct-type accesses which has helped code
readability.